manage users permissions

+++ resources/views/backend/blog/trash-table.blade.php
@@ -12,16 +12,18 @@
     @foreach($posts as $post)
         <tr>
             <td>
-                {!! Form::open(['style' => 'display:inline-block', 'method' => 'PUT', 'route' => ['backend.blog.restore', $post->id]]) !!}
-                <button title="Restore" class="btn btn-xs btn-default">
-                    <i class="fa fa-refresh"></i>
-                </button>
-                {!! Form::close() !!}
-                {!! Form::open(['style' => 'display:inline-block', 'method' => 'DELETE', 'route' => ['backend.blog.force-destroy', $post->id]]) !!}
-                <button title="Delete Permanent" type="submit" class="btn btn-xs btn-danger">
-                    <i class="fa fa-times"></i>
-                </button>
-                {!! Form::close() !!}
+                @if(check_user_permissions(request(), "Blog@forceDestroy", $post->id))
+                    {!! Form::open(['style' => 'display:inline-block', 'method' => 'PUT', 'route' => ['backend.blog.restore', $post->id]]) !!}
+                    <button title="Restore" class="btn btn-xs btn-default">
+                        <i class="fa fa-refresh"></i>
+                    </button>
+                    {!! Form::close() !!}
+                    {!! Form::open(['style' => 'display:inline-block', 'method' => 'DELETE', 'route' => ['backend.blog.force-destroy', $post->id]]) !!}
+                    <button title="Delete Permanent" type="submit" class="btn btn-xs btn-danger">
+                        <i class="fa fa-times"></i>
+                    </button>
+                    {!! Form::close() !!}
+                @endif
             </td>
             <td>{{$post->title}}</td>
             <td>{{$post->author->name}}</td>
